+++ src/chip8/opcode.rs
@@ -2,22 +2,22 @@
 pub struct Opcode {
     pub opcode: u16,
     pub code: u8,
-    pub x: u8,
-    pub y: u8,
+    pub x: usize,
+    pub y: usize,
     pub d: u8,
     pub kk: u8,
-    pub nnn: u16,
+    pub nnn: usize,
 }
 
 impl Opcode {
     pub fn new(byte1: u8, byte2: u8) -> Opcode {
         let opcode = make_opcode(byte1, byte2);
         let code = ((opcode & 0xF000) >> 12) as u8;
-        let x = ((opcode & 0x0F00) >> 8) as u8;
-        let y = ((opcode & 0x00F0) >> 4) as u8;
+        let x = ((opcode & 0x0F00) >> 8) as usize;
+        let y = ((opcode & 0x00F0) >> 4) as usize;
         let d = (opcode & 0x000F) as u8;
 
-        let nnn = opcode & 0x0FFF;
+        let nnn = (opcode & 0x0FFF) as usize;
         let kk: u8 = (opcode & 0x00FF) as u8;
 
         Opcode {
